feat: refactor, add sqlite db support
