+++ pkg/backend/persistent/history_event.go
@@ -2,7 +2,8 @@ package persistent
 
 import (
 	"context"
-	"github.com/tuannh982/simple-workflow-go/pkg/backend/psql/persistent/base"
+
+	"github.com/tuannh982/simple-workflow-go/pkg/backend/persistent/base"
 	"gorm.io/gorm"
 )
