+++ pkg/dataconverter/data_converter.go
@@ -1,6 +1,6 @@
 package dataconverter
 
-type DataConverter interface {
+type Codec interface {
 	Unmarshal(data []byte, v any) error
 	Marshal(v any) ([]byte, error)
 }
