+++ pkg/dataconverter/json_data_converter.go
@@ -4,7 +4,7 @@ import "encoding/json"
 
 type JsonDataConverter struct{}
 
-func NewJsonDataConverter() DataConverter {
+func NewJsonDataConverter() Codec {
 	return &JsonDataConverter{}
 }
